end test cours and permission add seed

+++ modules/studiobarg/RolePermission/Http/Controllers/RolePermissionsController.php
@@ -3,15 +3,18 @@
 namespace studiobarg\RolePermission\Http\Controllers;
 
 use App\Http\Controllers\Controller;
-use studiobarg\Category\Responses\AjaxResponse;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use studiobarg\Common\Responses\AjaxResponses;
 use studiobarg\RolePermission\Http\Requests\RoleRequest;
 use studiobarg\RolePermission\Http\Requests\RoleUpdateRequest;
+use studiobarg\RolePermission\Models\Role;
 use studiobarg\RolePermission\Repositories\PermissionRepo;
 use studiobarg\RolePermission\Repositories\RoleRepo;
 
 
 class RolePermissionsController extends Controller
 {
+    use AuthorizesRequests;
     private RoleRepo $roleRepo;
     private PermissionRepo $permissionRepo;
 
@@ -23,20 +26,24 @@ class RolePermissionsController extends Controller
 
     public function index()
     {
+        $this->authorize('index', Role::class);
         $roles = $this->roleRepo->all();
         $permissions = $this->permissionRepo->all();
 
         return view('RolePermissions::index', compact('roles', 'permissions'));
     }
 
+
     public function store(RoleRequest $request)
     {
+        $this->authorize('create', Role::class);
         $this->roleRepo->create($request);
-        return back();
+        return redirect(route('role-permissions.index'));
     }
 
     public function edit($roleId)
     {
+        $this->authorize('edit', Role::class);
         $role = $this->roleRepo->findById($roleId);
         $permissions = $this->permissionRepo->all();
         if (!$role) {
@@ -47,6 +54,7 @@ class RolePermissionsController extends Controller
 
     public function update($id, RoleUpdateRequest $request)
     {
+        $this->authorize('edit', Role::class);
         $role = $this->roleRepo->findById($id);
 
         $this->roleRepo->update($id, $request);
@@ -56,8 +64,9 @@ class RolePermissionsController extends Controller
 
     public function destroy($roleId)
     {
+        $this->authorize('delete', Role::class);
         $this->roleRepo->delete($roleId);
-        return AjaxResponse::successResponse();
+        return AjaxResponses::successResponse();
     }
 
 }
